renamings and restructuring
